Luxon Library Time and Date, Arrow functions, let and const
