--- modules/display-time
@@ -1,30 +1,0 @@
-// import { DateTime } from "./node_modules/luxon/src/luxon.js";
-// import { dateContainer } from "./querySelectors/date-container";
-
-const displayTime = () => {
-    const dateAndTime = new Date();
-    let completeDateAndTime = '';
-    completeDateAndTime += dateAndTime.toLocaleDateString();
-    completeDateAndTime += ',  ';
-    completeDateAndTime += dateAndTime.toLocaleTimeString();
-    completeDateAndTime = completeDateAndTime.replaceAll('/', ' ');
-    completeDateAndTime = completeDateAndTime.split(' ');
-    const month = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
-    completeDateAndTime[0] = month[completeDateAndTime[0] - 1];
-    if (completeDateAndTime[1] > 3) {
-      completeDateAndTime[1] = `${completeDateAndTime[1]}th`;
-    }
-    if (completeDateAndTime[1] === 3) {
-      completeDateAndTime[1] = `${completeDateAndTime[1]}rd`;
-    }
-    if (completeDateAndTime[1] === 2) {
-      completeDateAndTime[1] = `${completeDateAndTime[1]}nd`;
-    }
-    if (completeDateAndTime[1] === 1) {
-      completeDateAndTime[1] = `${completeDateAndTime[1]}st`;
-    }
-    completeDateAndTime = completeDateAndTime.join(' ');
-    return completeDateAndTime;
-}
-
-export { displayTime };
