--- modules/querySelectors/author-input
@@ -1,3 +1,0 @@
-const authorInput = document.querySelector('#author')
-
-export default authorInput;
