--- modules/querySelectors/submit-button
@@ -1,3 +1,0 @@
-const submitButton = document.querySelector('#submit-btn');
-
-export { submitButton as default };
