--- modules/querySelectors/title-input
@@ -1,3 +1,0 @@
-const titleInput = document.querySelector('#title')
-
-export default titleInput;
