+++ modules/sections.js
@@ -1,0 +1,4 @@
+
+export { listSection } from './list-section.js';
+export { addSection } from './add-section.js';
+export { contactSection } from './contact-section.js';
